Isolate controlled list manager into separate django app
